re-factor convert_polygons_to_sequences
