+++ darwin/exporter/formats/instance_mask.py
@@ -8,7 +8,7 @@ from PIL import Image
 from upolygon import draw_polygon
 
 import darwin.datatypes as dt
-from darwin.utils import get_progress_bar
+from darwin.utils import convert_polygons_to_sequences, get_progress_bar
 
 
 def generate_instance_id(instance_ids, length=8):
@@ -42,49 +42,6 @@ def export(annotation_files: Generator[dt.AnnotationFile, None, None], output_di
                 mask.save(masks_dir / f"{mask_id}.png")
                 f.write(f"{image_id},{mask_id},{cat}\n")
         print(f"Dataset format saved at {output_dir}")
-
-
-def convert_polygons_to_sequences(
-    polygons: List, height: Optional[int] = None, width: Optional[int] = None
-) -> List[np.ndarray]:
-    """
-    Converts a list of polygons, encoded as a list of dictionaries of into a list of nd.arrays
-    of coordinates.
-
-    Parameters
-    ----------
-    polygons: list
-        List of coordinates in the format [{x: x1, y:y1}, ..., {x: xn, y:yn}] or a list of them
-        as  [[{x: x1, y:y1}, ..., {x: xn, y:yn}], ..., [{x: x1, y:y1}, ..., {x: xn, y:yn}]].
-
-    Returns
-    -------
-    sequences: list[ndarray[float]]
-        List of arrays of coordinates in the format [[x1, y1, x2, y2, ..., xn, yn], ...,
-        [x1, y1, x2, y2, ..., xn, yn]]
-    """
-    if not polygons:
-        raise ValueError("No polygons provided")
-
-    # If there is a single polygon composing the instance then this is
-    # transformed to polygons = [[{x: x1, y:y1}, ..., {x: xn, y:yn}]]
-    if isinstance(polygons[0], dict):
-        polygons = [polygons]
-
-    if not isinstance(polygons[0], list) or not isinstance(polygons[0][0], dict):
-        raise ValueError("Unknown input format")
-
-    sequences = []
-    for polygon in polygons:
-        path = []
-        for point in polygon:
-            # Clip coordinates to the image size
-            x = max(min(point["x"], width - 1) if width else point["x"], 0)
-            y = max(min(point["y"], height - 1) if height else point["y"], 0)
-            path.append(round(x))
-            path.append(round(y))
-        sequences.append(path)
-    return sequences
 
 
 def convert_polygons_to_mask(polygons, height, width):
