+++ darwin/torch/dataset.py
@@ -6,7 +6,8 @@ import numpy as np
 from darwin.dataset import LocalDataset
 from darwin.dataset.utils import load_pil_image
 from darwin.torch.transforms import Compose, ConvertPolygonsToInstanceMasks, ConvertPolygonsToSegmentationMask
-from darwin.torch.utils import convert_polygons_to_sequences, polygon_area
+from darwin.torch.utils import polygon_area
+from darwin.utils import convert_polygons_to_sequences
 
 
 def get_dataset(
