+++ darwin/utils.py
@@ -64,6 +64,12 @@ def is_deprecated_project_dir(project_path: Path) -> bool:
     return (project_path / "annotations").exists() and (project_path / "images").exists()
 
 
+def get_progress_bar(array: List, description: Optional[str] = None):
+    pbar = tqdm(array)
+    pbar.set_description(desc=description, refresh=True)
+    return pbar
+
+
 def prompt(msg: str, default: Optional[str] = None) -> str:
     """Prompt the user on a CLI to input a message
 
@@ -209,7 +215,45 @@ def parse_darwin_annotation(annotation):
     return main_annotation
 
 
-def get_progress_bar(array: List, description: Optional[str] = None):
-    pbar = tqdm(array)
-    pbar.set_description(desc=description, refresh=True)
-    return pbar
+def convert_polygons_to_sequences(
+    polygons: List, height: Optional[int] = None, width: Optional[int] = None
+) -> List[np.ndarray]:
+    """
+    Converts a list of polygons, encoded as a list of dictionaries of into a list of nd.arrays
+    of coordinates.
+
+    Parameters
+    ----------
+    polygons: list
+        List of coordinates in the format [{x: x1, y:y1}, ..., {x: xn, y:yn}] or a list of them
+        as  [[{x: x1, y:y1}, ..., {x: xn, y:yn}], ..., [{x: x1, y:y1}, ..., {x: xn, y:yn}]].
+
+    Returns
+    -------
+    sequences: list[ndarray[float]]
+        List of arrays of coordinates in the format [[x1, y1, x2, y2, ..., xn, yn], ...,
+        [x1, y1, x2, y2, ..., xn, yn]]
+    """
+    if not polygons:
+        raise ValueError("No polygons provided")
+    # If there is a single polygon composing the instance then this is
+    # transformed to polygons = [[{x: x1, y:y1}, ..., {x: xn, y:yn}]]
+    if isinstance(polygons[0], dict):
+        polygons = [polygons]
+
+    if not isinstance(polygons[0], list) or not isinstance(polygons[0][0], dict):
+        raise ValueError("Unknown input format")
+
+    sequences = []
+    for polygon in polygons:
+        path = []
+        for point in polygon:
+            # Clip coordinates to the image size
+            x = max(min(point["x"], width - 1) if width else point["x"], 0)
+            y = max(min(point["y"], height - 1) if height else point["y"], 0)
+            path.append(round(x))
+            path.append(round(y))
+        sequences.append(np.array(path))
+    return sequences
+
+
